tests: moved health report related benchmark to cluster module
