+++ src/v/cluster/tests/CMakeLists.txt
@@ -184,3 +184,11 @@ rp_test(
   ARGS "-- -c 1"
   LABELS cluster
 )
+
+rp_test(
+  BENCHMARK_TEST
+  BINARY_NAME health_monitor
+  SOURCES health_monitor_bench.cc
+  LIBRARIES Seastar::seastar_perf_testing v::cluster
+  LABELS cluster
+)
