+++ src/v/serde/test/CMakeLists.txt
@@ -10,7 +10,7 @@ rp_test(
   BENCHMARK_TEST
   BINARY_NAME serde
   SOURCES bench.cc
-  LIBRARIES Seastar::seastar_perf_testing v::serde v::cluster
+  LIBRARIES Seastar::seastar_perf_testing v::serde
   LABELS serde
 )
